:truck: refactor folder structure
