+++ src/app/plan/[slug]/components/server/ProgressDot.tsx
