+++ src/app/plan/[slug]/components/server/TripInfoCard.tsx
