+++ src/components/shared/Providers/Providers.tsx
@@ -1,7 +1,7 @@
-import AxiosProvider from '@/app/components/provider/AxiosProvider'
-import ResponsiveHeightProvider from '@/app/components/provider/ResponsiveHeightProvider'
-import QueryClientProvider from '@/app/components/provider/TanstackQueryProvider'
-import ToastifyProvider from '@/app/components/provider/ToastifyProvider'
+import AxiosProvider from '@/components/provider/AxiosProvider'
+import ResponsiveHeightProvider from '@/components/provider/ResponsiveHeightProvider'
+import QueryClientProvider from '@/components/provider/TanstackQueryProvider'
+import ToastifyProvider from '@/components/provider/ToastifyProvider'
 
 interface Props {
   children: React.ReactNode
