+++ src/components/shared/SNB/SNB.constants.ts
